Use pointers in page guard and add moves
Page guards require the use of a move constructor in particular for the
case of using std::optional with the page guards. Additionally to
support the transition to use move constructors the underlying handles
to bot the frame headers and buffer pool managers have been changed
to shared_pointers to prevent mishaps when deconstructing moved
objects.