Update interface to use raw pointers
Right now I'm using smart pointers (unique pointer) for storing the
frame data so I don't forget to clean it up and call .get(). This might
be a code smell that I should look into in the future. Also fixed a
bug in the code that tried to acquire a lock twice (once on construction
and afterward explicitly with a .lock() call).