Fix race condition in acquiring lock in page guard constructor
There was a race condition leading to a deadlock in the constructor of
the page guards. This deadlock occurred as the frame header mutex was
being held while trying to construct another page guard from inside the buffer
pool manager. But the frame header mutex could not be released as it required
the lock to the buffer pool manager. A polling approach was taken to fix this
issue.