fix: fix blocking when creating logger
Unit tests were failing as the loggers for the disk manager and page
buffer manager were not being created properly. In particular, the
initial creation of the logger was fine, however, when reopening the
logger (even after creating an entirely new instance of the managers and
deconstructing the old one) there was a blocking call to create the
logger as one already existed with that name in the spdlog internal
state. This was fixed with calling spdlog::get to fetch the logger if
one existed previously.