Support potential failure while waiting for page guard creation
The current implementation of the page buffer pool doesn't support any
handling for errors when waiting for a page. Specifically, the case
where I/O fails this could lead to a complete deadlock in the program.
Continuously retrying the I/O operation isn't very good so an
alternative approach is now being taken of returning an optional value
for a page guard for the wait functions. The main different being that
that only the case of I/O failure will std::nullopt be returned.

Furthermore, to add some extra safety the original wait condition has
been changed to check if a) the page is already in the case, or b) there
is an evictable page to remove to make space for the page. This is
almost complete but some small changes must be added for the starting
case when the LRU replacement hasn't been fully loaded with page data
and the frames have been loaded.