Increment evictable count on frame registration
There was a small bug with the register of frames. When the page buffer
manager was creating frames and then registering them so that the frames
in the buffer could be tracked using the replacement policy. However on
frame registration the count for evictable frames was not being
incremented making it seem as if there were no evictable frames on
startup.