Move AddAccessor call to page buffer manager
Originally the call to accessor call for pages was placed inside of the page guard. This is a bit excessive and could possibly create some performance issues in the case that a page guard is being created on a page that is already loaded as it will hold the lock on the page buffer manager for far longer than is needed. The lock ends immediately after updating the count.