Begin stubbing out what the slotted page implementation is going to look like
This is the start of the implemenation for the slotted page format. At a high level the buffer pool manager (hopefully) should be refactored to pass these logical pages instead of the guards themselves. TBD whether or not the page guard will just be placed directly inside of the pages themeselves or they will be combined into a single object. This will require a pretty big refactor especially in the lower level tests I assume.