Update interface to make use of Page classes
With the creation of the Page classes for slotted page formatting, the
page buffer manager no longer uses the page guards. Also, the interface
was slightly changed for adding and removing accessors, to take page
id's since all frame information should be all local to the page buffer
manager and should not concern any other components or classes.