fix: re-order class members to fix race condition
There was a very odd race conditioning happening every so often not even
running multiple commands concurrently to the disk scheduler by just
creating a new page. It seemed like due to the order of the members in
the class definition the worker thread was being made right away leading
to it running referencing potentially initialized members in the class.