Add configuration for logging and database file naming
Currently there is a very large amount of coupling on the design behind logging forcing file logging. This isn't great in case of test or when uses just want a singular database file. As such a configuration will be created to pass along for each of the different loggers for the different components of the db