Slightly updating interface
Reducing down the size of the interface a bit to make it more intuitive to use. It's clear now while adding the checksum functionality that a factory method might make sense to implement so that I can return an error if for instance there was corruption and the page is in an invalid state (i.e. checksum is invalid)