Implement BasePage class
Since initially the read-only page instance was a parent class and the mutage page inherrited there were some odd interactions since, although it was almost an is-a relationship it wasn't. The main pain point being the way that the lock for either class was held since they required a unique and shared locked respectively this kind of led to a strange design. The BasePage essentially implements everything the read only page had except for locking. For this reason the BasePage should never be initialized by itself otherwise there will be race conditions.