Don't return optional for writing or reading page
As I think of transitioning to removing the disk scheduler altogether it doesn't really make any more sense to return a std::optional for whne trying to acquire a page. I think throwing an exception in this case (likely at the disk manager level) makes a lot more sense than having the error propogate upwards through all of the layers. This fix is just a small little hack. The interface is correct but the implementation is all wrong at the moment. Going to need to refactor after implementing the slotted pages